Organise classes into one package
