+++ src/main/java/BottleOpener/Deadline.java
@@ -1,3 +1,5 @@
+package BottleOpener;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
@@ -7,7 +9,7 @@ import java.time.temporal.ChronoField;
 /**
  * Represents a task with a deadline.
  * <p>
- * A Deadline task includes a description, a status indicating whether it is completed,
+ * A BottleOpener.Deadline task includes a description, a status indicating whether it is completed,
  * and a deadline represented as a {@link LocalDateTime}.
  * </p>
  */
@@ -38,14 +40,14 @@ public class Deadline extends Task {
     public LocalDateTime deadline;
 
     /**
-     * Constructs a new Deadline task with the specified description and deadline.
+     * Constructs a new BottleOpener.Deadline task with the specified description and deadline.
      * The task is initially marked as not done.
      * <p>
      * The deadline is parsed from the given string using the defined {@link #FORMATTER}.
      * If the deadline is not in a valid format, the current date and time will be used.
      * </p>
      *
-     * @param description The description of the Deadline task.
+     * @param description The description of the BottleOpener.Deadline task.
      * @param deadline    The deadline of the task as a string.
      */
     public Deadline(String description, String deadline) {
@@ -59,13 +61,13 @@ public class Deadline extends Task {
     }
 
     /**
-     * Constructs a new Deadline task with the specified description, status, and deadline.
+     * Constructs a new BottleOpener.Deadline task with the specified description, status, and deadline.
      * <p>
      * The deadline is parsed from the given string using the defined {@link #FORMATTER}.
      * If the deadline is not in a valid format, the current date and time will be used.
      * </p>
      *
-     * @param description The description of the Deadline task.
+     * @param description The description of the BottleOpener.Deadline task.
      * @param status      The completion status of the task; {@code true} if the task is done, {@code false} otherwise.
      * @param deadline    The deadline of the task as a string.
      */
@@ -80,9 +82,9 @@ public class Deadline extends Task {
     }
 
     /**
-     * Constructs a new Deadline task with the specified description, status, and deadline.
+     * Constructs a new BottleOpener.Deadline task with the specified description, status, and deadline.
      *
-     * @param description The description of the Deadline task.
+     * @param description The description of the BottleOpener.Deadline task.
      * @param status      The completion status of the task; {@code true} if the task is done, {@code false} otherwise.
      * @param deadline    The deadline of the task as a {@link LocalDateTime} object.
      */
@@ -102,7 +104,7 @@ public class Deadline extends Task {
     /**
      * Returns the type of the task.
      *
-     * @return "D", representing that this task is a Deadline.
+     * @return "D", representing that this task is a BottleOpener.Deadline.
      */
     public String getType() {
         return "D";
@@ -118,7 +120,7 @@ public class Deadline extends Task {
     }
 
     /**
-     * Returns a string representation of the Deadline task, including its type, status icon, description, and deadline.
+     * Returns a string representation of the BottleOpener.Deadline task, including its type, status icon, description, and deadline.
      *
      * @return A string in the format "[type] [status icon] description (by: deadline)".
      */
