+++ src/main/java/BottleOpener/Storage.java
@@ -1,17 +1,19 @@
+package BottleOpener;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 /**
- * The {@code Storage} class handles loading and saving tasks from/to a file.
+ * The {@code BottleOpener.Storage} class handles loading and saving tasks from/to a file.
  * It ensures that task data is persisted between sessions.
  */
 public class Storage {
     public Path path;
 
     /**
-     * Constructs a {@code Storage} object with the specified file path.
+     * Constructs a {@code BottleOpener.Storage} object with the specified file path.
      *
      * @param filepath The file path where tasks will be loaded from and saved to.
      */
@@ -20,12 +22,12 @@ public class Storage {
     }
 
     /**
-     * Loads tasks from the specified file into the given {@code Tasklist}.
+     * Loads tasks from the specified file into the given {@code BottleOpener.Tasklist}.
      * If the file does not exist, it is created. The method parses the file content and
-     * converts each line into a {@code Task} object, adding it to the list.
+     * converts each line into a {@code BottleOpener.Task} object, adding it to the list.
      *
-     * @param tasks The {@code Tasklist} to load the data into.
-     * @return The updated {@code Tasklist} after loading the data.
+     * @param tasks The {@code BottleOpener.Tasklist} to load the data into.
+     * @return The updated {@code BottleOpener.Tasklist} after loading the data.
      */
     public Tasklist load(Tasklist tasks) {
         try {
@@ -64,7 +66,7 @@ public class Storage {
      * Saves the current list of tasks to the specified file.
      * The tasks are serialized into a specific format and written to the file.
      *
-     * @param tasks The {@code Tasklist} to be saved.
+     * @param tasks The {@code BottleOpener.Tasklist} to be saved.
      */
     public void save(Tasklist tasks) {
         String outputFile = "";
