+++ src/main/java/BottleOpener/Tasklist.java
@@ -1,7 +1,9 @@
+package BottleOpener;
+
 import java.util.ArrayList;
 
 /**
- * The {@code Tasklist} class manages a list of tasks, allowing operations such as adding,
+ * The {@code BottleOpener.Tasklist} class manages a list of tasks, allowing operations such as adding,
  * deleting, marking, and unmarking tasks. It provides functionality to interact with the
  * task list and retrieve information about the tasks.
  */
@@ -10,7 +12,7 @@ public class Tasklist {
     public ArrayList<Task> tasklist;
 
     /**
-     * Constructs an empty {@code Tasklist} with an initial index of 0.
+     * Constructs an empty {@code BottleOpener.Tasklist} with an initial index of 0.
      */
     public Tasklist() {
         this.index = 0;
@@ -18,7 +20,7 @@ public class Tasklist {
     }
 
     /**
-     * Constructs a {@code Tasklist} with a predefined list of tasks.
+     * Constructs a {@code BottleOpener.Tasklist} with a predefined list of tasks.
      *
      * @param tasklist The list of tasks to be managed.
      */
