+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 -----------------------------------
-Hello! I'm BottleOpener.
+Hello! I'm BottleOpener.BottleOpener.
 What can I do for you?
 -----------------------------------
